A2A/language-learning-academy/pyproject.toml
[project]
name = "language-learning-academy"
version = "0.1.0"
description = "Language Learning Academy agent for vocabulary, grammar, and conversation practice"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "sse-starlette>=2.3.5",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
    "structlog>=24.1.0",
    "langdetect>=1.0.9",
    "streamlit>=1.28.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
]

[project.scripts]
language-learning-academy = "language_learning_academy.server.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
